rename worldstate to state
